+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   utils: Added currency conversions from any denominations to smallest (Atto)
+-   utils: Added currency conversions from any denominations to largest
 
 ### Changed